#5 account and auth code divided from 'boards'
