+++ authentication/views.py
@@ -1,3 +1,30 @@
-from django.shortcuts import render
+from rest_framework import generics, viewsets, permissions, filters
+from django_filters.rest_framework import DjangoFilterBackend
+from .serializers import RegisterSerializer, BusinessSerializer
+from .models import Business, BusinessMember
+from authentication.permissions import IsBusinessAdmin
 
-# Create your views here.
+
+# 사용자 등록을 위한 뷰
+class RegisterView(generics.CreateAPIView):
+    """
+    새로운 BusinessMember (사용자)을 등록하는 API 엔드포인트.
+    모든 사용자가 접근할 수 있습니다.
+    """
+    queryset = BusinessMember.objects.all()
+    permission_classes = (permissions.AllowAny,)
+    serializer_class = RegisterSerializer
+
+# 비즈니스 관리용 뷰셋
+class BusinessViewSet(viewsets.ModelViewSet):
+    """
+    Business 모델의 CRUD (생성, 조회, 수정, 삭제) 기능을 제공하는 뷰셋.
+    오직 비즈니스 관리자(Admin)만 접근할 수 있습니다.
+    """
+    queryset = Business.objects.all()
+    serializer_class = BusinessSerializer
+    permission_classes = [permissions.IsAuthenticated, IsBusinessAdmin]
+    filter_backends = [DjangoFilterBackend, filters.SearchFilter, filters.OrderingFilter]
+    filterset_fields = ['name']
+    search_fields = ['name', 'address', 'phone_number', 'website']
+    ordering_fields = ['name']
