+++ boards/urls.py
@@ -2,14 +2,12 @@
 
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-from .views import PostViewSet, CommentViewSet, BusinessViewSet, RegisterView
+from .views import PostViewSet, CommentViewSet
 
 router = DefaultRouter()
 router.register('posts', PostViewSet, basename='post')
 router.register('comments', CommentViewSet, basename='comment')
-router.register('businesses', BusinessViewSet, basename='business')
 
 urlpatterns = [
     path('', include(router.urls)),
-    path('auth/register/', RegisterView.as_view(), name='auth_register'),
 ]
